feat: approve to permit2 in helper hook as well
when first deposit happens through the helper hook, we know also approve permit2 contract for that asset as well. This is because euler vault contracts try to pull using permit2 first by default and if it fails, it tries to do a simple transferFrom. If we have approve permit2 as well, it saves gas. It increases gas for the first depositor but that is ok sinces it is just one time cost